+++ src/components/OurPartners/OurPartner.helpers.ts
@@ -1,0 +1,19 @@
+import companyBlack from "../../../public/assets/images/companyBlack.svg";
+import companyYellow from "../../../public/assets/images/companyYellow.svg";
+import companyPink from "../../../public/assets/images/companyPink.svg";
+import companyGreen from "../../../public/assets/images/companyGreen.svg";
+import companyPurple from "../../../public/assets/images/companyPurple.svg";
+import companyOrange from "../../../public/assets/images/companyOrange.svg";
+
+export const yourCompanyLogos = [
+  companyBlack,
+  companyYellow,
+  companyBlack,
+  companyPink,
+  companyBlack,
+  companyGreen,
+  companyBlack,
+  companyPurple,
+  companyBlack,
+  companyOrange,
+];
