+++ src/mockData/mockData.ts
@@ -1,9 +1,4 @@
-const companyBlack = "/assets/images/companyBlack.svg";
-const companyYellow = "/assets/images/companyYellow.svg";
-const companyPink = "/assets/images/companyPink.svg";
-const companyGreen = "/assets/images/companyGreen.svg";
-const companyPurple = "/assets/images/companyPurple.svg";
-const companyOrange = "/assets/images/companyOrange.svg";
+
 // -----------------------------------------------------------
 
 
@@ -42,18 +37,6 @@ const confidence2 = '/assets/icon/confid.svg'
 
 
 
-export const yourCompanyLogos = [
-  companyBlack,
-  companyYellow,
-  companyBlack,
-  companyPink,
-  companyBlack,
-  companyGreen,
-  companyBlack,
-  companyPurple,
-  companyBlack,
-  companyOrange,
-];
 
 
 export const mockdataIcon = [
